Bug 1393800 - Have mochitests expecting crashes wait for the crashes to be recorded before clean up; r=mconley
This patch includes a bunch of somewhat related fixes, these are:

- Ensuring that when a mochitest calls SimpleTest.expectChildProcessCrash()
  the harness will wait for the crashes to be recorded before deleting the
  dump files. This involves a message round-trip between the content and
  parent process so to minimize its performance impact on all the non-crashing
  tests it is done only when required.
- As an additional optimization, the SimpleTest harness will not send a
  message to the content process anymore whenever it receives an
  ipc:content-shutdown event, instead it does it only for abnormal shutdowns.
- Manually fixing remaining mochitests causing crashes to wait for crashes to
  be recorded before finishing and deleting the dump files.
- Modifying BrowserTestUtils.crashBrowser() so that it optionally does not
  delete the dump files, this is useful for tests that submit their dumps and
  thus delete them on their own.


MozReview-Commit-ID: 4SLJ8BjJ18n

UltraBlame original commit: b9a113790a15d9480f4426a9ee6fbdd40b83386d

+++ testing/mochitest/BrowserTestUtils/BrowserTestUtils.jsm
@@ -1006,7 +1006,10 @@ this.BrowserTestUtils = {
 
 
 
-  async crashBrowser(browser, shouldShowTabCrashPage=true) {
+
+
+  async crashBrowser(browser, shouldShowTabCrashPage=true,
+                     shouldClearMinidumps=true) {
     let extra = {};
     let KeyValueParser = {};
     if (AppConstants.MOZ_CRASHREPORTER) {
@@ -1107,8 +1110,10 @@ this.BrowserTestUtils = {
               }
             }
 
-            removeFile(minidumpDirectory, dumpID + '.dmp');
-            removeFile(minidumpDirectory, dumpID + '.extra');
+            if (shouldClearMinidumps) {
+              removeFile(minidumpDirectory, dumpID + '.dmp');
+              removeFile(minidumpDirectory, dumpID + '.extra');
+            }
           });
         }
